+++ kanu/docgpt.py
@@ -17,7 +17,7 @@ from langchain.document_loaders import (
     CSVLoader,
 )
 
-from .utils import Tooltip
+from .utils import Tooltip, Settings
 
 DOCUMENT_LOADERS = {
     ".txt": (TextLoader, {"encoding": "utf8"}),
@@ -29,11 +29,13 @@ DOCUMENT_LOADERS = {
 
 class DocGPT:
     def __init__(self, kanu, openai_key, model, temperature, prompt):
+        self.name = "DocGPT"
         self.kanu = kanu
         self.model = model
         self.temperature = temperature
         self.prompt = prompt
         os.environ["OPENAI_API_KEY"] = openai_key
+        self.settings = Settings(self)
 
     def run(self):
         self.kanu.container.pack_forget()
@@ -103,22 +105,26 @@ class DocGPT:
         self.kanu.container = tk.Frame(self.kanu.root)
         self.kanu.container.pack()
         l = tk.Label(self.kanu.container, text="DocGPT")
-        l.grid(row=0, column=0, columnspan=3)    
+        l.grid(row=0, column=0, columnspan=4)    
         self.session = tk.Text(self.kanu.container, width=70, height=20)
-        self.session.grid(row=1, column=0, columnspan=3)
-        e = tk.Entry(self.kanu.container, width=54)
-        e.grid(row=2, column=0, columnspan=3)
-        b = tk.Button(self.kanu.container, text="Send", command=lambda: self.send_message(e))
+        self.session.grid(row=1, column=0, columnspan=4)
+        self.session.tag_config("user", background=self.settings.user_background_color.get(), foreground=self.settings.user_foreground_color.get())
+        self.session.tag_config("bot", background=self.settings.bot_background_color.get(), foreground=self.settings.user_foreground_color.get())
+        user_input = tk.Entry(self.kanu.container, width=54)
+        user_input.grid(row=2, column=0, columnspan=4)
+        b = tk.Button(self.kanu.container, text="Send", command=lambda: self.send_message(user_input))
         b.grid(row=3, column=0)
         b = tk.Button(self.kanu.container, text="Clear", command=lambda: self.clear_session())
         b.grid(row=3, column=1)
         b = tk.Button(self.kanu.container, text="Go back", command=lambda: self.run())
         b.grid(row=3, column=2)
+        b = tk.Button(self.kanu.container, text="Settings", command=lambda: self.settings.page())
+        b.grid(row=3, column=3)
 
     def send_message(self, entry):
-        self.session.insert(tk.END, "You: " + entry.get() + "\n")
+        self.session.insert(tk.END, "You: " + entry.get() + "\n", "user")
         response = self.qa(entry.get())["answer"]
-        self.session.insert(tk.END, "Bot: " + response + "\n")
+        self.session.insert(tk.END, "Bot: " + response + "\n", "bot")
         entry.delete(0, tk.END)
 
     def go_with_option1(self):