+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.6.0 (in development)
 * Enable users to chat with .csv documents.
+* Enable users to customize chat settings (e.g. font size and background color).
 
 ## 0.5.0 (2023-06-08)
 * Enable users to ask follow-up questions when chatting with documents.